Add files via upload
Currently, the robot can deal with either attitude control or position control. But it's messing up managing both transitional and rotational commands at the same time. the "low_level_control.py" gives an example of control on rotationl commands, and the example of transitional control can be exhibited by uncommenting line#104 and line#107