Add files via upload
To integrate this control algorithm with path planning, only the ref() function needs to be modified.Basically, it will recieve an array in the form of [position, linear velocity, attitude, angular velocity] . And the attitude is expressed with quaternoin. It's going to be 3(pos) + 3(vel) + 4(att) + 3(ang vel) =13 elements in the input array.